Add initial Core Data support

+++ Healthub/Repositories/RepositoriesProtocol/TherapyRepositoryProtocol.swift
@@ -10,7 +10,7 @@ import Foundation
 protocol TherapyRepositoryProtocol{
     func getDrugList(query: String,  completionHandler: @escaping ([Drug]?, Error?) -> Void)
     func getAll(completionHandler: @escaping ([Therapy]?, Error?) -> Void)
-    func createTherapy(drug_ids: [Int], duration: String, name: String, comment: String, completionHandler: @escaping (Bool?, Error?) -> Void)
+    func createTherapy(drug_ids: [Int16], duration: String, name: String, comment: String, completionHandler: @escaping (Bool?, Error?) -> Void)
     /*func addTherapy()
     func deleteTherapy()*/
